Implement random pay days
Running the simulation, we would see all companies and people follow
the same monthly trend: companies all inc money, people all dec
money. Then at the end of the month, when everyone gets paid, so
everyone spikes in the other direction all at once. This made it
difficult to find trends in the plot, so to smooth out this
movement, I assigned every company a random pay day in the month.

While this did smooth out the plot, it didn't as much as I expected
it to. Results of 5 and 10 year simulations show that the spikes
start in the beginning, when everyone starts with the same amount,
but smooths out as these amounts diverge over time. However, the
trends stayed the same, leading me to believe that this doesn't
affect the trend; though it makes the plot easier to read.