Reverted back to same pay day, and sampled plots for easier readability
Since randomizing the pay day didn't actually change the trends
in the results, I chose to revert back so everyone has the same
pay day, to minimize variables. But address the spikes in the plots
from before, I redid the plots so that they sample the first day
of each month. This basically picks one point in up-down cycle to
represent the month, so it's a smoother graph showing the trend