Change spending rate to a fraction of total money, rather than monthly income
When spending is a fraction of monthly income, people only spend up to 25%
more than they take in each month. This means that if someone has a large
buffer of savings, the probability of it getting used is very low. (Think
about this - if the range is 50/50 spend vs save, the prob of saving 9
months in a row is 0.5^9 ~ 2/1000. So in the small economy with 1000 people,
we're virtually guaranteed that at least one person will build a buffer for
almost a year, which we see happening in the graph. Even if they dip after
that, the chance of them losing all of that money is equally small.)

We don't see this trend happening with companies though, because
surplus cash in companies is *checked* by hiring additional employees. So
we see ups and downs. So we needed a way to check people who have surplus
cash. This approach works because as wealth increases, the probability of
spending more than your income (which is fixed) increases. So you're less
likely to net positive each month, which serves as a check.

The resulting plots show an equilibrium where the wealth distribution
among people and companies stabilizes, and unemployment stabilizes when
companies monthly revenue can afford to pay their employees, on average -
which is around 25-30% unemployment.