Fix bug in layoff logic
When optimizing the logic, I forgot to divide income by 12, so
companies were way overcalculating how much to take off their
payroll when laying off an employee. The test for layoffs didn't
catch this because it didn't use a big enough test case, so I
adjusted the test to catch this case. I also changed the income
variable to be monthly income, to avoid having to divide by 12
every time you want to use it. (Though the config still specifies
income as annual income)