Combining the web & websocket servers by using flask-sockets
The flask-sockets module allows a flask server to respond to
websocket endpoints, so I decided to use that instead. The websocket
endpoint is currently very simple, reading 2 numbers and sending
back a Fibonacci sequence starting with those two numbers. This also
meant I was able to revert back to the default Python version
installed on the vagrant machine (3.6), so I did that too.