Optimize updates to charts during simulation
The server used to send all all results over all the days in
each update to the client, and the client would process these.
This means the amount of data sent increases over the course of
the simulation, slowing it down. So now the server sends an update
each day, and only sends the results from that day. The client
has been updated accordingly. So now it runs at constant speed.