Change new_money parameters to one value, rename "stimulus"
The distribution of "new money" was not really useful. The goal is
to model stimulus checks, and a fraction of income is closer to
what's done in reality. For instance, in the US' first relief
package for covid, everyone got a fixed check; unemployed people
get an amount proportional to their income. Similarly, in France,
people get at least 80% of their income.

Ref: https://www.economist.com/united-states/2020/05/16/inefficiencies-bedevil-americas-response-to-soaring-joblessness