Move results callback to the beginning of each day, from end of day
It doesn't really matter when (start vs end) we choose to report the
results, as long as it's done every day. But it just turns out to be
easier to reason about when it's at the beginning, because on
every 30th day, this captures the state right after payroll was paid.