Fix bug: daily spending calculated incorrectly
Each person is supposed to pick a fraction of their money to spend
each month, and spend that amount. But the amount was calculated each
day as a fraction of their *current* total of money. So when we
subtract that amount, the next day, the amount is a fraction of
*what's left*. This actually means people are spending less than
they should, because the amount decreases each day. I fixed this
(I think), by setting the amount to spend each day in advance,
and explicitly spending that amount each day.